PaymentOverlayHandler Update
The “bringToFrontAndDeliver” function has been updated to include a new parameter, “clearTop.” This parameter allows the function to clear the top element of the activity stack before bringing the "WebViewPaymentActivity" to the front. This update addresses the use case of destroying the CustomTabActivity after the login flow, specifically when using Custom Chrome Tab.