[APPC-5012] MainActivity Update
The MainActivity has been updated to encapsulate the logic for managing the snackBar's message.
Upon receiving a new intent, a helper function checks for the presence of a snackBar message and stores it in a Channel (within the ViewModel) for later be consumed by a fragment hosted within the MainActivity.