[APPC-5108] Implementation of SetUpCurrencyActivity
This activity is responsible for receiving the deeplink and creating a wallet, if necessary, to facilitate the setup of the user’s local currency.

This implementation will be relevant to ensure the correct currency in cloud gaming before the user makes a purchase within the game. This will contribute to a better user experience.